refactor: 즐겨찾기 버튼 컴포넌트화
